move index.js to lib/server.js
